Adding Assignment 2 php files
Assignment instructions: 'Construct an HTML page with a short article
about a particular piece of technology that had a profound effect on
humankind. Some good examples might be the steam locomotive, radio or
even the humble desktop mouse. The page should include at least one
image with an appropriate credit to the photographer or illustrator. The
page will also feature a form, allowing a visitor to purchase this piece
of technology.'